Turn off setup mode after FlightController goes away
It is impossible for firmware to figure out on it's own if FlightController
connected or not. So FC sent a command at exit to turn setup mode off.
Guess how reliable that was (actually, it was mostly working).

For some time now, FC polls the device for status ~3 times per second.
So we can have an idle timeout (set to 2000 ticks(=milliseconds) now) after
which we can consider FlightController gone.

Firmware will not restart on that timeout - so if FC crashes, you won't lose
your uncommitted changes. It should restart scan module - to prevent spam
in case said configuration is not sane. But apply_config() seems to crash
the firmware somehow, so it is not done - it just clears the status reg bit