"dma_core" -> "cortex".
Because "dma_core" ees a stoopid name. I just needed something that's not "tmk_core" at the time.
And cortex is the part of a brain that does zee thinking (most CommonSense varieties also have "reflexes" part, which are part of the particular project)

I considered "core" - but "core/core.h" is a bit too much.